refactor(detection): extract shared integration-project vocabulary; drop headless-scope
Replace the headless-specific detection module with a chooser-agnostic
one, @lib/detection/integration-projects: scan + classify (moved from
self-driving's detect-agentic, now with the recommended flag),
chooseProject (auto-choose policy for unattended runs), and
resolveProjectDir (containment, previously duplicated as self-driving's
integrationDir). Self-driving consumes it through a thin adapter — its
picker is just a different chooser over the same report — and a future
interactive basic-integration picker takes the same report with
recommended as the default selection.

The headless plumbing (early auth, flag gate, logging, analytics tags)
moves into the basic-integration program, whose policy it is.

Generated-By: PostHog Code
Task-Id: 06aaaf67-5aba-497e-bd9e-0220e2417442
